diet-analysis-functions/requirements.txt
azure-functions
azure-storage-blob
pandas
matplotlib
